:sparkles: 이벤트 감사 로그(Audit Log) 테이블 구현
