+++ apps/commerce-streamer/src/main/java/com/loopers/domain/EventLogRepository.java
@@ -1,0 +1,10 @@
+package com.loopers.domain;
+
+import java.util.List;
+
+public interface EventLogRepository {
+
+	void save(EventLog eventLog);
+	
+	void saveAll(List<EventLog> eventLogs);
+}
