+++ apps/commerce-streamer/src/main/java/com/loopers/infrastructure/EventLogJpaRepository.java
@@ -1,0 +1,9 @@
+package com.loopers.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.loopers.domain.EventLog;
+
+public interface EventLogJpaRepository extends JpaRepository<EventLog, Long> {
+
+}
